Fix all 57 golangci-lint warnings (closes #19)
Resolve all pre-existing lint issues that CI skips via --only-new-issues:
- errcheck (26): Add `_ =` for intentionally ignored cleanup/close errors
- gosec (21): Add nolint directives with justifications for false positives,
  tighten test file permissions to 0o600/0o750
- unused (8): Move Linux-only ONNX download code to download_linux.go,
  remove unused mockMicChecker from tui_test.go
- staticcheck (2): Use http.Status constants, simplify TrimPrefix
