+++ internal/transcriber/openai_test.go
@@ -16,36 +16,36 @@ func TestOpenAITranscribe(t *testing.T) {
 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if r.URL.Path != "/v1/audio/transcriptions" {
 			t.Errorf("unexpected path: %s", r.URL.Path)
-			http.Error(w, "not found", 404)
+			http.Error(w, "not found", http.StatusNotFound)
 			return
 		}
 		if r.Method != http.MethodPost {
 			t.Errorf("unexpected method: %s", r.Method)
-			http.Error(w, "method not allowed", 405)
+			http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
 			return
 		}
 
-		err := r.ParseMultipartForm(10 << 20)
+		err := r.ParseMultipartForm(10 << 20) //nolint:gosec // test code with bounded input
 		if err != nil {
 			t.Errorf("parse multipart: %v", err)
-			http.Error(w, "bad request", 400)
+			http.Error(w, "bad request", http.StatusBadRequest)
 			return
 		}
 
-		receivedModel = r.FormValue("model")
-		receivedFormat = r.FormValue("response_format")
+		receivedModel = r.FormValue("model")       //nolint:gosec // test code
+		receivedFormat = r.FormValue("response_format") //nolint:gosec // test code
 
 		file, _, err := r.FormFile("file")
 		if err != nil {
 			t.Errorf("get form file: %v", err)
-			http.Error(w, "bad request", 400)
+			http.Error(w, "bad request", http.StatusBadRequest)
 			return
 		}
-		defer file.Close()
+		defer func() { _ = file.Close() }()
 		receivedFileData, _ = io.ReadAll(file)
 
 		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("  Hello world  "))
+		_, _ = w.Write([]byte("  Hello world  "))
 	}))
 	defer server.Close()
 
@@ -72,7 +72,7 @@ func TestOpenAITranscribe(t *testing.T) {
 
 func TestOpenAITranscribeError(t *testing.T) {
 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		http.Error(w, "model not found", 404)
+		http.Error(w, "model not found", http.StatusNotFound)
 	}))
 	defer server.Close()
