+++ internal/tui/tui_test.go
@@ -32,19 +32,6 @@ func (m *mockLevelSampler) AudioLevel() float64 {
 	return m.level
 }
 
-type mockMicChecker struct {
-	available bool
-	name      string
-}
-
-func (m *mockMicChecker) MicAvailable() bool {
-	return m.available
-}
-
-func (m *mockMicChecker) MicName() string {
-	return m.name
-}
-
 type mockPostProcessor struct {
 	result string
 	err    error
